separate legend item in a component
